server/pyproject.toml
[project]
name = "mondash-server"
version = "0.1.0"
description = "MonDash - ESP32-S3 Monitoring Dashboard Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pysnmp-lextudio>=6.1.0",
    "apscheduler>=3.10.0",
    "pillow>=10.0.0",
    "ffmpeg-python>=0.2.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[project.scripts]
mondash = "app.main:run"
